referral_app/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
python-dotenv==1.0.0
celery==5.3.1
psycopg2-binary==2.9.7